WIP: Genesis parallelism
